[KAIZEN-0] Lager egen utils-mappe og renamer tidsstempel til endret_dato
