+++ src/main/java/no/nav/fo/veilarbarena/utils/DateUtils.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbarena;
+package no.nav.fo.veilarbarena.utils;
 
 import javax.xml.datatype.XMLGregorianCalendar;
 import java.time.LocalDate;
