+++ src/main/java/no/nav/fo/veilarbarena/utils/MigrationUtils.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbarena;
+package no.nav.fo.veilarbarena.utils;
 
 import org.flywaydb.core.Flyway;
