+++ src/main/java/no/nav/fo/veilarbarena/utils/ServiceUtils.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbarena;
+package no.nav.fo.veilarbarena.utils;
 
 import no.nav.sbl.dialogarena.types.Pingable;
